requirements.txt
google-genai
python-dotenv==1.0.0
numpy 
flask==2.3.3
pydantic==2.5.0
